move common parts
